+++ db/db.go
@@ -9,15 +9,17 @@ import (
 	"hmruntime/logger"
 	"hmruntime/manifest"
 	"hmruntime/metrics"
+	"hmruntime/utils"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
 )
 
-var globalInferenceWriter *inferenceWriter
+var globalInferenceWriter *inferenceWriter = &inferenceWriter{}
 
 const batchSize = 100
 const chanSize = 10000
+const inferencesTable = "inferences"
 
 const inferenceRefresherInterval = 5 * time.Second
 
@@ -30,8 +32,8 @@ type inferenceWriter struct {
 
 type inferenceHistory struct {
 	model  manifest.Model
-	input  string
-	output string
+	input  any
+	output any
 	start  time.Time
 	end    time.Time
 }
@@ -50,7 +52,7 @@ func Stop() {
 	globalInferenceWriter.dbpool.Close()
 }
 
-func WriteInferenceHistory(model manifest.Model, input, output string, start, end time.Time) {
+func WriteInferenceHistory(model manifest.Model, input, output any, start, end time.Time) {
 	globalInferenceWriter.Write(inferenceHistory{
 		model:  model,
 		input:  input,
@@ -79,7 +81,7 @@ func (w *inferenceWriter) worker(ctx context.Context) {
 			batchIndex = 0
 			timer.Reset(inferenceRefresherInterval)
 		case <-w.quit:
-			WriteInferenceHistoryToDB(ctx, batch[:batchSize])
+			WriteInferenceHistoryToDB(ctx, batch[:batchIndex])
 			close(w.done)
 			return
 		}
@@ -90,15 +92,19 @@ func WriteInferenceHistoryToDB(ctx context.Context, batch []inferenceHistory) {
 	if len(batch) == 0 {
 		return
 	}
-	table := os.Getenv("NAMESPACE")
-	if table == "" {
-		table = "local_instance"
-	}
 	err := WithTx(ctx, func(tx pgx.Tx) error {
 		b := &pgx.Batch{}
 		for _, data := range batch {
-			query := fmt.Sprintf("INSERT INTO %s (model_hash, input, output, started_at, duration_ms) VALUES ($1, $2, $3, $4, $5)", table)
-			args := []interface{}{data.model.Hash(), data.input, data.output, data.start, data.end.Sub(data.start).Milliseconds()}
+			input, err := utils.JsonSerialize(data.input)
+			if err != nil {
+				return err
+			}
+			output, err := utils.JsonSerialize(data.output)
+			if err != nil {
+				return err
+			}
+			query := fmt.Sprintf("INSERT INTO %s (id, model_hash, input, output, started_at, duration_ms) VALUES ($1, $2, $3, $4, $5, $6)", inferencesTable)
+			args := []any{utils.GeneratUUID(), data.model.Hash(), input, output, data.start, data.end.Sub(data.start).Milliseconds()}
 			b.Queue(query, args...)
 		}
 
@@ -117,13 +123,13 @@ func WriteInferenceHistoryToDB(ctx context.Context, batch []inferenceHistory) {
 
 	if err != nil {
 		// Handle error
-		logger.Error(ctx).Err(err).Msg("Error writing to inference history database")
+		logger.Err(ctx, err).Msg("Error writing to inference history database")
 	}
 
 }
 
 func Initialize(ctx context.Context) {
-	connStr := os.Getenv("HYPERMODE_CLUSTER_DB")
+	connStr := os.Getenv("HYPERMODE_METADATA_DB")
 
 	var err error
 	tempDBPool, err := pgxpool.New(ctx, connStr)
@@ -152,11 +158,7 @@ func WithTx(ctx context.Context, fn func(pgx.Tx) error) error {
 	if err != nil {
 		return err
 	}
-	defer func() {
-		if err := tx.Rollback(ctx); err != nil {
-			logger.Error(ctx).Err(err).Msg("Failed to rollback transaction")
-		}
-	}()
+	defer tx.Rollback(ctx)
 
 	err = fn(tx)
 	if err != nil {