+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## UNRELEASED
 
--  Add Model Inference History to runtime [#186](https://github.com/gohypermode/runtime/pull/186)
+- Add Model Inference History to runtime [#186](https://github.com/gohypermode/runtime/pull/186)
 - Pass auth headers correctly when invoking a GraphQL API [#196](https://github.com/gohypermode/runtime/pull/196)
 
 ## 2024-05-13 - Version 0.7.0